EFFF9.c

+++ include/sf33rd/Source/Game/EFFF9.h
@@ -6,6 +6,6 @@
 
 void effect_F9_move(WORK_Other *ewk);
 void effect_F9_init(s16 END_PL_NO);
-s32 Rewrite_End_Message(u8 mes_no);
+s32 Rewrite_End_Message(u16 mes_no);
 
 #endif
